moved INotificationSender to Application layer
